# CompProgEdu/WebApp/src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CompProgEdu API",
    "version": "v1"
  },
  "paths": {
    "/api/assignments/by-id/{id}": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "operationId": "GetById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AssignmentDetailDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/assignments/by-course-id/{courseId}": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "operationId": "GetAllAssignmentsByCourseId",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AssignmentSummaryDtoListPaginatedResponse"
            }
          }
        }
      }
    },
    "/api/assignments/student/{courseId}": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "operationId": "GetAllStudentAssignmentsByCourseId",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AssignmentSummaryDtoListValidateableResponse"
            }
          }
        }
      }
    },
    "/api/assignments": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "operationId": "Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateAssignmentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AssignmentGetDtoValidateableResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Assignments"
        ],
        "operationId": "Update",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateAssignmentByIdRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AssignmentDetailDtoValidateableResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Assignments"
        ],
        "operationId": "Delete",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AssignmentGetDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/assignments/assignment-submission": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "operationId": "SubmitAssignment",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "assignmentId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "code",
            "type": "string"
          },
          {
            "in": "query",
            "name": "language",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "submission",
            "type": "string",
            "format": "binary"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SubmissionResultValidateableResponse"
            }
          }
        }
      }
    },
    "/api/authenticate/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "Login",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/LoginResponseValidateableResponse"
            }
          }
        }
      }
    },
    "/api/authenticate/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "Register",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserGetDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/authenticate/register-student": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "RegisterStudent",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterStudentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserGetDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/authenticate/register-instructor": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "RegisterInstructor",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterInstructorRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserGetDtoValidateableResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "CourseGetDto": {
      "required": [
        "section",
        "title"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "section": {
          "type": "string"
        },
        "instructorId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "Instructor": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        }
      }
    },
...
    "StudentCourse": {
      "type": "object",
      "properties": {
        "studentId": {
          "format": "int32",
          "type": "integer"
        },
        "student": {
          "$ref": "#/definitions/Student"
        },
        "courseId": {
          "format": "int32",
          "type": "integer"
        },
        "course": {
          "$ref": "#/definitions/Course"
        }
      }
    },
    "Course": {
      "required": [
        "section",
        "title"
      ],
      "type": "object",
      "properties": {
        "instructor": {
          "$ref": "#/definitions/Instructor"
        },
        "studentCourses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StudentCourse"
          }
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "section": {
          "type": "string"
        },
        "instructorId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "MethodTestCase": {
      "required": [
        "output",
        "paramInputs"
      ],
      "type": "object",
      "properties": {
        "assignment": {
          "$ref": "#/definitions/Assignment"
        },
        "input": {
          "type": "string",
          "readOnly": true
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "returnType": {
          "type": "string"
        },
        "methodTestInjectable": {
          "type": "string"
        },
        "assignmentId": {
          "format": "int32",
          "type": "integer"
        },
        "paramInputs": {
          "type": "string"
        },
        "output": {
          "type": "string"
        },
        "hint": {
          "type": "string"
        },
        "pointValue": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Assignment": {
      "required": [
        "allowedLanguages",
        "assignmentInstructions",
        "assignmentName",
        "exampleInput",
        "exampleOutput",
        "totalPointValue"
      ],
      "type": "object",
      "properties": {
        "course": {
          "$ref": "#/definitions/Course"
        },
        "desiredOutputs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DesiredOutput"
          }
        },
        "methodTestCases": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MethodTestCase"
          }
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentName": {
          "type": "string"
        },
        "allowedLanguages": {
          "type": "string"
        },
        "assignmentInstructions": {
          "type": "string"
        },
        "exampleInput": {
          "type": "string"
        },
        "exampleOutput": {
          "type": "string"
        },
        "totalPointValue": {
          "format": "int32",
          "type": "integer"
        },
        "dueDate": {
          "format": "date-time",
          "type": "string"
        },
        "visibilityDate": {
          "format": "date-time",
          "type": "string"
        },
        "assignmentSolutionFileName": {
          "type": "string"
        },
        "courseId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "DesiredOutput": {
      "required": [
        "input",
        "output"
      ],
      "type": "object",
      "properties": {
        "assignment": {
          "$ref": "#/definitions/Assignment"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentId": {
          "format": "int32",
          "type": "integer"
        },
        "input": {
          "type": "string"
        },
        "output": {
          "type": "string"
        },
        "pointValue": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "AssignmentDetailDto": {
      "required": [
        "allowedLanguages",
        "assignmentInstructions",
        "assignmentName",
        "exampleInput",
        "exampleOutput",
        "totalPointValue"
      ],
      "type": "object",
      "properties": {
        "courseDto": {
          "$ref": "#/definitions/CourseGetDto"
        },
        "desiredOutputs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DesiredOutput"
          }
        },
        "methodTestCases": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MethodTestCase"
          }
        },
        "totalPointsAssigned": {
          "format": "int32",
          "type": "integer"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentName": {
          "type": "string"
        },
        "allowedLanguages": {
          "type": "string"
        },
        "assignmentInstructions": {
          "type": "string"
        },
        "exampleInput": {
          "type": "string"
        },
        "exampleOutput": {
          "type": "string"
        },
        "totalPointValue": {
          "format": "int32",
          "type": "integer"
        },
        "dueDate": {
          "format": "date-time",
          "type": "string"
        },
        "visibilityDate": {
          "format": "date-time",
          "type": "string"
        },
        "assignmentSolutionFileName": {
          "type": "string"
        },
        "courseId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "fieldName": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "AssignmentDetailDtoValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/AssignmentDetailDto"
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "AssignmentSummaryDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "courseId": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentName": {
          "type": "string"
        },
        "assignmentInstructions": {
          "type": "string"
        },
        "totalPointValue": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentSolution": {
          "type": "string"
        },
        "dueDate": {
          "type": "string"
        },
        "visibilityDate": {
          "type": "string"
        }
      }
    },
    "AssignmentSummaryDtoListPaginatedResponse": {
      "type": "object",
      "properties": {
        "totalCount": {
          "format": "int32",
          "type": "integer"
        },
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AssignmentSummaryDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "AssignmentSummaryDtoListValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AssignmentSummaryDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "CreateAssignmentRequest": {
      "required": [
        "allowedLanguages",
        "assignmentInstructions",
        "assignmentName",
        "exampleInput",
        "exampleOutput",
        "totalPointValue"
      ],
      "type": "object",
      "properties": {
        "assignmentName": {
          "type": "string"
        },
        "allowedLanguages": {
          "type": "string"
        },
        "assignmentInstructions": {
          "type": "string"
        },
        "exampleInput": {
          "type": "string"
        },
        "exampleOutput": {
          "type": "string"
        },
        "totalPointValue": {
          "format": "int32",
          "type": "integer"
        },
        "dueDate": {
          "format": "date-time",
          "type": "string"
        },
        "visibilityDate": {
          "format": "date-time",
          "type": "string"
        },
        "assignmentSolutionFileName": {
          "type": "string"
        },
        "courseId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "AssignmentGetDto": {
      "required": [
        "allowedLanguages",
        "assignmentInstructions",
        "assignmentName",
        "exampleInput",
        "exampleOutput",
        "totalPointValue"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentName": {
          "type": "string"
        },
        "allowedLanguages": {
          "type": "string"
        },
        "assignmentInstructions": {
          "type": "string"
        },
        "exampleInput": {
          "type": "string"
        },
        "exampleOutput": {
          "type": "string"
        },
        "totalPointValue": {
          "format": "int32",
          "type": "integer"
        },
        "dueDate": {
          "format": "date-time",
          "type": "string"
        },
        "visibilityDate": {
          "format": "date-time",
          "type": "string"
        },
        "assignmentSolutionFileName": {
          "type": "string"
        },
        "courseId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "AssignmentGetDtoValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/AssignmentGetDto"
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "UpdateAssignmentByIdRequest": {
      "required": [
        "allowedLanguages",
        "assignmentInstructions",
        "assignmentName",
        "exampleInput",
        "exampleOutput",
        "totalPointValue"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentName": {
          "type": "string"
        },
        "allowedLanguages": {
          "type": "string"
        },
        "assignmentInstructions": {
          "type": "string"
        },
        "exampleInput": {
          "type": "string"
        },
        "exampleOutput": {
          "type": "string"
        },
        "totalPointValue": {
          "format": "int32",
          "type": "integer"
        },
        "dueDate": {
          "format": "date-time",
          "type": "string"
        },
        "visibilityDate": {
          "format": "date-time",
          "type": "string"
        },
        "assignmentSolutionFileName": {
          "type": "string"
        },
        "courseId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "MethodTestCheckDto": {
      "type": "object",
      "properties": {
        "passed": {
          "type": "boolean"
        },
        "hint": {
          "type": "string"
        }
      }
    },
    "SubmissionResult": {
      "type": "object",
      "properties": {
        "totalScore": {
          "format": "int32",
          "type": "integer"
        },
        "outputChecks": {
          "type": "array",
          "items": {
            "type": "boolean"
          }
        },
        "methodTestChecks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MethodTestCheckDto"
          }
        }
      }
    },
    "SubmissionResultValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/SubmissionResult"
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "stayLoggedIn": {
          "type": "boolean"
        }
      }
    },
    "UserGetDto": {
      "required": [
        "emailAddress",
        "firstName",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "emailAddress": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "firstName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "lastName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
    "LoginResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/UserGetDto"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "LoginResponseValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/LoginResponse"
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "RegisterRequest": {
      "required": [
        "emailAddress",
        "firstName",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        },
        "emailAddress": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "firstName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "lastName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
    "UserGetDtoValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/UserGetDto"
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "RegisterStudentRequest": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "studentSchoolNumber": {
          "type": "string"
        },
        "changedPassword": {
          "type": "boolean"
        }
      }
    },
    "RegisterInstructorRequest": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer {token}"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}